REGRESSION(254760@main): mobile version of apple.com/iphone-14 doesn't display https://bugs.webkit.org/show_bug.cgi?id=246122 rdar://100814764
Reviewed by NOBODY (OOPS!).

The page doesn't work because it tries to initialize a DOMMatrix with a value of `translate(-50%, -50%)` which is invalid given it contains a percentage value.
Before 254760@main, the transform used to compute to `none` because it is on an element with no renderer, and `new DOMMatrix("none")` initializes fine, which gets the page working again.

Test: LayoutTests/fast/css/transform-in-non-rendered-subtree-computed-style.html

* Source/WebCore/css/ComputedStyleExtractor.cpp:
(WebCore::computedTransform):
* LayoutTests/fast/css/transform-in-non-rendered-subtree-computed-style.html: Added.
* LayoutTests/fast/css/transform-in-non-rendered-subtree-computed-style-expected.txt: Added.