[JSC] Array literal parsing should lex StringLiteral in the end of parsing https://bugs.webkit.org/show_bug.cgi?id=245066 rdar://99815328
Reviewed by Mark Lam.

public class field can introduce a case that string literal after the array literal can have
a semantically meaningful content. Previously, we only checked it is string, and we didn't need
to have a string content itself for this case,

    [ ] "Hello"

But now, if it is used with class public field, then we need to know the content of the string to
emit a syntax error based on that.

    class A {
        g = []
        "constructor"() { }
    }

This patch drops array literal parsing's DontBuildStrings flag to lex the string content.

* JSTests/stress/array-literal-lexing-for-class-public-fields.js: Added.
(shouldBe):
(main.a.prototype.g.string_appeared_here):
(main.a):
(main):
* Source/JavaScriptCore/parser/Parser.cpp:
(JSC::Parser<LexerType>::parseArrayLiteral):

Canonical link: https://commits.webkit.org/255212@main