KeyframeEffect::setAnimation should clear blending key frames for an old CSSAnimation https://bugs.webkit.org/show_bug.cgi?id=246244
Reviewed by Antoine Quint.

KeyframeEffect::propertyAffectingKeyframeResolutionDidChange performs a switch on m_blendingKeyframesSource but this one is no longer valid if the animation has changed.
For CSSTransition this is a no-op and for WebAnimation this is just a call to clearBlendingKeyframes(), so wrong behavior can only happen with CSSAnimation.

* Source/WebCore/animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::setAnimation): Clear blending key frames if the old animation was a CSSAnimation to prevent any misbehavior in later calls to propertyAffectingKeyframeResolutionDidChange.

Canonical link: https://commits.webkit.org/255308@main