"display" property value of computed style for audio element without controls is "inline" (not "none")
"display" property value of computed style for audio element without controls is "inline" (not "none")

https://bugs.webkit.org/show_bug.cgi?id=88105

Reviewed by Devin Rousso.

This is to align Webkit behavior with Blink / Chrome and Gecko / Firefox and Web-Specifications:

https://html.spec.whatwg.org/#embedded-content-rendering-rules

"When an audio element is not exposing a user interface, the user agent is expected
to force its 'display' property to compute to 'none', irrespective of CSS rules."

When an audio element has no controls attribute, it should not be visible on the page.
Before this patch, when we removed the "controls" attribute, an empty div of the size of the controls remained.

Tests from Patch Authored by Silvia Pfeiffer.

* Source/WebCore/css/html.css: Update UA stylesheet where the audio controls are not visible
* LayoutTests/media/audio-controls-hidden-display-none.html: Added Test Cases
* LayoutTests/media/audio-controls-hidden-display-none-expected.txt: Added Test Case Expectations

Canonical link: https://commits.webkit.org/255528@main