WebXR should respond to changes in frame size https://bugs.webkit.org/show_bug.cgi?id=246496 <rdar://problem/101151986>
Reviewed by Antoine Quint.

The size of the framebuffer provided by the compositor may change during the
life of a session. Detect this and respond to it.

* Source/WebCore/Modules/webxr/WebXROpaqueFramebuffer.cpp:
(WebCore::WebXROpaqueFramebuffer::startFrame): Look at the size of the incoming frame.
If it has changed, re-run the setupFramebuffer logic.
(WebCore::WebXROpaqueFramebuffer::setupFramebuffer): Make sure we delete existing
WebGL objects if we create new ones.

Canonical link: https://commits.webkit.org/255543@main